Add import/no-restricted-paths to imports rules
This rule was added by 1.10.0. I am setting it to be disabled because it
is very project-specific, so it doesn't make sense to enable it here.

https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-restricted-paths.md

I found the sections of this file a little confusing, so I didn't know
where to put it. I think it would be better to simply alphabetize the
rules but I'll leave that for another time.